Add setup script and update docs
